+++ src/hooks/GJLevelScoreCell.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include <Geode/modify/GJLevelScoreCell.hpp>
+
+using namespace geode::prelude;
+
+class $modify(MyGJLevelScoreCell, GJLevelScoreCell) {
+    void loadFromScore(GJUserScore* p0);
+};