+++ src/hooks/GJRequestCell.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include <Geode/modify/GJRequestCell.hpp>
+
+using namespace geode::prelude;
+
+class $modify(MyGJRequestCell, GJRequestCell) {
+    void loadFromScore(GJUserScore* p0);
+};