+++ src/hooks/GJScoreCell.cpp
@@ -1,12 +1,20 @@
 #include "GJScoreCell.hpp"
 #include "../FineOutline.hpp"
+#include <hiimjasmine00.user_data_api/include/Events.hpp>
+#include <hiimjasmine00.user_data_api/include/UserDataAPI.hpp>
 
-void MyGJScoreCell::loadFromScore(GJUserScore* p0) {
-    GJScoreCell::loadFromScore(p0);
+void MyGJScoreCell::loadFromScore(GJUserScore* userScore) {
+    GJScoreCell::loadFromScore(userScore);
 
-    if (p0->m_accountID != GJAccountManager::get()->m_accountID) return;
+    auto player = typeinfo_cast<SimplePlayer*>(getChildByIDRecursive("player-icon"));
+    if (!player) return;
 
-    if (auto player = typeinfo_cast<SimplePlayer*>(getChildByIDRecursive("player-icon"))) {
+    if (userScore->m_accountID == GJAccountManager::get()->m_accountID) {
         alpha::fine_outline::impl::setOutlineColor(player, alpha::fine_outline::impl::getColor(alpha::fine_outline::PlayerIcon::ONE));
+    }
+    else {
+        user_data::handleScoreCell(this, [this, player] (GJUserScore* score) {
+            alpha::fine_outline::impl::setOutlineColorOnline("player-1-color", score, player);
+        });
     }
 }