+++ src/hooks/GJUserCell.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include <Geode/modify/GJUserCell.hpp>
+
+using namespace geode::prelude;
+
+class $modify(MyGJUserCell, GJUserCell) {
+    void loadFromScore(GJUserScore* p0);
+};